♻️ A run leaves one artifact, one header and one report (#31)
* Simplify run artifacts and frozen embedding caches

* Represent artifact retention with one optional reason

* docs: record the Task_5 artifact contract and the cited canonical config versions


* Reject existing run output names before writing artifacts

* Create run artifacts without overwrite and update the CI smoke

* docs: the smoke recipe produces its own baseline before the diff


---------

+++ crates/cmem-eval-runner/src/diff.rs
@@ -195,13 +195,13 @@ mod tests {
     use super::*;
     use serde_json::{Value, json};
 
-    fn row(run_id: &str, latency_ms: u128, rank: usize) -> PerQuestionResult {
+    fn row(run_id: &str, latency_ms: u64, rank: usize) -> PerQuestionResult {
         row_with(run_id, latency_ms, rank, |_| {})
     }
 
     fn row_with(
         run_id: &str,
-        latency_ms: u128,
+        latency_ms: u64,
         rank: usize,
         edit: impl FnOnce(&mut Value),
     ) -> PerQuestionResult {
